applied fixes from 2.2.1

+++ src/main/java/com/rae/crowns/content/fields/util/client/LocalPhysicData.java
@@ -49,7 +49,7 @@ public class LocalPhysicData {
         int localY = pos.getY() & 15;
         int localZ = pos.getZ() & 15;
 
-        return layer.get(localX, localY, localZ);
+        return layer.get((short) localX, (short) localY, (short) localZ);
     }
 
     public static @NotNull Set<SectionPos> getTickingSections() {
